+++ src/main/java/com/cgi/dao/RoleDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.Role;
 
-public interface RoleDao extends GenericDao<Role, Integer> {
+public interface RoleDao extends GenericDao<Role, Long> {
 
 }
