+++ src/main/java/com/cgi/dao/UserDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.User;
 
-public interface UserDao extends GenericDao<User, Integer> {
+public interface UserDao extends GenericDao<User, Long> {
 
 }
